fix(WBC): fix control during aerial state
It seems the WBC is unstable when the dog is completely in air.

Specifically, when a joint is accelerating, it causes the torso
to accelerate in the other direction. Thus, the torso acceleration
will cause the joint to apply more torque. However, in fact the
torso is not accelerated so fast, and the additional torque makes
the joint unstable.

I have to reduce the foot kp and kd in order to stable it.
However, the true reason why it is unstable is still unknown.