feat(Half Flip): control the dog to perform half flip
Finally! After all the tinkering, tuning and bug-fixing,
the controller is able to control the dog to perform continuous
side flip.
However, this is not exactly what I expected. Because in this
controller, I manually planned a trajectory, and the controller
simply tried to follow it as close as possible, without any sort
of modification to the trajectory. So the controller can't tell
whether the trajectory is physically feasible.
Later, I may try to add a feature to allow controller generate
its own trajectory.